Ruby 2.5 included in Appraisal
Rails 6 and Rails 4 demand a different version of Ruby so we had to
include code to control ruby version.

The code implies that the gemfiles by appraisal will be created and used
only when certain ruby versions are being userd. Thus, the Rails 4.0
gemfiles will olny be used, and the tests run, when in a version below
ruby 2.4.